requirements.txt
Flask>=2.3.0
Flask-CORS>=4.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-experimental>=0.3.0
langchain-chroma>=0.2.0
openai>=1.10.0
chromadb
tiktoken>=0.7.0
markdown>=3.5.0
gunicorn>=21.0.0
aiohttp>=3.9.0
markupsafe>=2.1.0
jinja2>=3.1.0
werkzeug>=2.3.0
numpy
pandas
rank_bm25